requirements.txt
<<<<<<< HEAD
appdirs==1.4.4
asgiref==3.2.7
astroid==2.4.2
autopep8==1.5.3
certifi==2020.4.5.1
colorama==0.4.3
distlib==0.3.0
Django==3.0.6
django-filter==2.3.0
filelock==3.0.12
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.4
pandas==1.0.3
Pillow==7.1.2
pipenv==2018.11.26
psycopg2==2.8.5
pycodestyle==2.6.0
pylint==2.5.3
pylint-django==2.0.15
pylint-plugin-utils==0.6
python-dateutil==2.8.1
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
virtualenv==20.0.20
virtualenv-clone==0.5.4
wrapt==1.12.1
=======
appdirs==1.4.4
asgiref==3.2.7
astroid==2.4.2
autopep8==1.5.3
certifi==2020.4.5.1
colorama==0.4.3
distlib==0.3.0
Django==3.0.6
django-filter==2.3.0
filelock==3.0.12
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.4
pandas==1.0.3
Pillow==7.1.2
pipenv==2018.11.26
psycopg2==2.8.5
pycodestyle==2.6.0
pylint==2.5.3
pylint-django==2.0.15
pylint-plugin-utils==0.6
python-dateutil==2.8.1
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
virtualenv==20.0.20
virtualenv-clone==0.5.4
wrapt==1.12.1
>>>>>>> 7511be836ccd1300326d827fd7d821c5e95ba3b4
